requirements.txt
Flask==2.3.3
openai==1.3.0
joblib==1.3.2
python-dotenv==1.0.0</content>
<parameter name="filePath">c:\Users\Win 0\OneDrive\Pictures\Documents\artifical intelligence\Source code\AI-PROJECT\requirements.txt